feat: adding authorixation checks

+++ backend/src/modules/posts/postService.ts
@@ -2,10 +2,13 @@ import { AppError } from "../../errors/AppError.js";
 import {prisma} from "../../lib/prisma.js";
 import {CreatePostInput,createPostSchema} from "./postSchema.js";
 import {deleteImage} from "../../utils/deleteImage.js"
+import { userInfo } from "node:os";
 
-export const getPostsService = async(perPage:number,currentPage:number)=>{
+export const getPostsService = async(userId:number,perPage:number,currentPage:number)=>{
     const posts = await prisma.post.findMany(
     {
+        
+        where:{creatorId:userId},
         include:
         {
             creator:
@@ -30,12 +33,12 @@ export const getPostsService = async(perPage:number,currentPage:number)=>{
     };
 
 };
-export const postPostService = async(data:CreatePostInput,imageFileName:string)=>{;
+export const postPostService = async(data:CreatePostInput,imageFileName:string,userId:number)=>{;
         const post = await prisma.post.create({data:{
             title: data.title,
             imageUrl: `/images/${imageFileName}`,
             content: data.content,
-            creatorId:1,
+            creatorId:userId,
         }});
         return post;
 }
@@ -56,13 +59,14 @@ export const getPostService = async(postId:number)=>{
 }
 
 export const updatePostService = async(
+    userId:number,
     postId:number,
     updatedTitle:string,
     updatedContent:string,
     updatedImageUrl?:string)=>{
     const post = await prisma.post.findUnique({
         where:{id:postId},
-        select:{id:true,imageUrl:true}
+        select:{id:true,imageUrl:true,creatorId:true}
     });
     //check whether post exist or not
     if(!post){throw new AppError("Couldn't find the post.",404)}
@@ -74,8 +78,12 @@ export const updatePostService = async(
         title:updatedTitle,
         content:updatedContent
     };
+    
     if(updatedImageUrl !== undefined){
         data.imageUrl = updatedImageUrl;
+    }
+    if(post.creatorId !== userId){
+        throw new AppError("This post doesn't belong to this user!",403);
     }
     const updatedPost = await prisma.post.update({
         where:{id:postId},
@@ -94,9 +102,10 @@ export const updatePostService = async(
     return updatedPost;
 
 }
-export const deletePostService = async(postId:number)=>{
-    const existingPost = await prisma.post.findUnique({where:{id:postId},select:{id:true,imageUrl:true}});
+export const deletePostService = async(userId:number,postId:number)=>{
+    const existingPost = await prisma.post.findUnique({where:{id:postId},select:{id:true,imageUrl:true,creatorId:true}});
     if(!existingPost){throw new AppError("post not found.",404);}
+    if(existingPost.creatorId !== userId){ throw new AppError("This post doesn't belong to this user!",403);}
     await prisma.post.delete({where:{id:postId}});
     if(existingPost.imageUrl){
         await deleteImage(existingPost.imageUrl);
